+++ DebugUtils/Source/DebugDraw.cpp
@@ -54,14 +54,14 @@ unsigned int duIntToCol(int i, int a)
 	return duRGBA(r*63,g*63,b*63,a);
 }
 
-void duIntToCol(int i, float* col)
+void duIntToCol(int i, double* col)
 {
 	int	r = bit(i, 0) + bit(i, 3) * 2 + 1;
 	int	g = bit(i, 1) + bit(i, 4) * 2 + 1;
 	int	b = bit(i, 2) + bit(i, 5) * 2 + 1;
-	col[0] = 1 - r*63.0f/255.0f;
-	col[1] = 1 - g*63.0f/255.0f;
-	col[2] = 1 - b*63.0f/255.0f;
+	col[0] = 1 - r*63.0/255.0;
+	col[1] = 1 - g*63.0/255.0;
+	col[2] = 1 - b*63.0/255.0;
 }
 
 void duCalcBoxColors(unsigned int* colors, unsigned int colTop, unsigned int colSide)
@@ -76,8 +76,8 @@ void duCalcBoxColors(unsigned int* colors, unsigned int colTop, unsigned int col
 	colors[5] = duMultCol(colSide, 217);
 }
 
-void duDebugDrawCylinderWire(struct duDebugDraw* dd, float minx, float miny, float minz,
-							 float maxx, float maxy, float maxz, unsigned int col, const float lineWidth)
+void duDebugDrawCylinderWire(struct duDebugDraw* dd, double minx, double miny, double minz,
+							 double maxx, double maxy, double maxz, unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 	
@@ -86,8 +86,8 @@ void duDebugDrawCylinderWire(struct duDebugDraw* dd, float minx, float miny, flo
 	dd->end();
 }
 
-void duDebugDrawBoxWire(struct duDebugDraw* dd, float minx, float miny, float minz,
-						float maxx, float maxy, float maxz, unsigned int col, const float lineWidth)
+void duDebugDrawBoxWire(struct duDebugDraw* dd, double minx, double miny, double minz,
+						double maxx, double maxy, double maxz, unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 	
@@ -96,9 +96,9 @@ void duDebugDrawBoxWire(struct duDebugDraw* dd, float minx, float miny, float mi
 	dd->end();
 }
 
-void duDebugDrawArc(struct duDebugDraw* dd, const float x0, const float y0, const float z0,
-					const float x1, const float y1, const float z1, const float h,
-					const float as0, const float as1, unsigned int col, const float lineWidth)
+void duDebugDrawArc(struct duDebugDraw* dd, const double x0, const double y0, const double z0,
+					const double x1, const double y1, const double z1, const double h,
+					const double as0, const double as1, unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 	
@@ -107,9 +107,9 @@ void duDebugDrawArc(struct duDebugDraw* dd, const float x0, const float y0, cons
 	dd->end();
 }
 
-void duDebugDrawArrow(struct duDebugDraw* dd, const float x0, const float y0, const float z0,
-					  const float x1, const float y1, const float z1,
-					  const float as0, const float as1, unsigned int col, const float lineWidth)
+void duDebugDrawArrow(struct duDebugDraw* dd, const double x0, const double y0, const double z0,
+					  const double x1, const double y1, const double z1,
+					  const double as0, const double as1, unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 	
@@ -118,8 +118,8 @@ void duDebugDrawArrow(struct duDebugDraw* dd, const float x0, const float y0, co
 	dd->end();
 }
 
-void duDebugDrawCircle(struct duDebugDraw* dd, const float x, const float y, const float z,
-					   const float r, unsigned int col, const float lineWidth)
+void duDebugDrawCircle(struct duDebugDraw* dd, const double x, const double y, const double z,
+					   const double r, unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 	
@@ -128,8 +128,8 @@ void duDebugDrawCircle(struct duDebugDraw* dd, const float x, const float y, con
 	dd->end();
 }
 
-void duDebugDrawCross(struct duDebugDraw* dd, const float x, const float y, const float z,
-					  const float size, unsigned int col, const float lineWidth)
+void duDebugDrawCross(struct duDebugDraw* dd, const double x, const double y, const double z,
+					  const double size, unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 	
@@ -138,8 +138,8 @@ void duDebugDrawCross(struct duDebugDraw* dd, const float x, const float y, cons
 	dd->end();
 }
 
-void duDebugDrawBox(struct duDebugDraw* dd, float minx, float miny, float minz,
-					float maxx, float maxy, float maxz, const unsigned int* fcol)
+void duDebugDrawBox(struct duDebugDraw* dd, double minx, double miny, double minz,
+					double maxx, double maxy, double maxz, const unsigned int* fcol)
 {
 	if (!dd) return;
 	
@@ -148,8 +148,8 @@ void duDebugDrawBox(struct duDebugDraw* dd, float minx, float miny, float minz,
 	dd->end();
 }
 
-void duDebugDrawCylinder(struct duDebugDraw* dd, float minx, float miny, float minz,
-						 float maxx, float maxy, float maxz, unsigned int col)
+void duDebugDrawCylinder(struct duDebugDraw* dd, double minx, double miny, double minz,
+						 double maxx, double maxy, double maxz, unsigned int col)
 {
 	if (!dd) return;
 	
@@ -158,9 +158,9 @@ void duDebugDrawCylinder(struct duDebugDraw* dd, float minx, float miny, float m
 	dd->end();
 }
 
-void duDebugDrawGridXZ(struct duDebugDraw* dd, const float ox, const float oy, const float oz,
-					   const int w, const int h, const float size,
-					   const unsigned int col, const float lineWidth)
+void duDebugDrawGridXZ(struct duDebugDraw* dd, const double ox, const double oy, const double oz,
+					   const int w, const int h, const double size,
+					   const unsigned int col, const double lineWidth)
 {
 	if (!dd) return;
 
@@ -179,29 +179,29 @@ void duDebugDrawGridXZ(struct duDebugDraw* dd, const float ox, const float oy, c
 }
 		 
 
-void duAppendCylinderWire(struct duDebugDraw* dd, float minx, float miny, float minz,
-						  float maxx, float maxy, float maxz, unsigned int col)
+void duAppendCylinderWire(struct duDebugDraw* dd, double minx, double miny, double minz,
+						  double maxx, double maxy, double maxz, unsigned int col)
 {
 	if (!dd) return;
 
 	static const int NUM_SEG = 16;
-	static float dir[NUM_SEG*2];
+	static double dir[NUM_SEG*2];
 	static bool init = false;
 	if (!init)
 	{
 		init = true;
 		for (int i = 0; i < NUM_SEG; ++i)
 		{
-			const float a = (float)i/(float)NUM_SEG*DU_PI*2;
+			const double a = (double)i/(double)NUM_SEG*DU_PI*2;
 			dir[i*2] = dtMathCosf(a);
 			dir[i*2+1] = dtMathSinf(a);
 		}
 	}
 	
-	const float cx = (maxx + minx)/2;
-	const float cz = (maxz + minz)/2;
-	const float rx = (maxx - minx)/2;
-	const float rz = (maxz - minz)/2;
+	const double cx = (maxx + minx)/2;
+	const double cz = (maxz + minz)/2;
+	const double rx = (maxx - minx)/2;
+	const double rz = (maxz - minz)/2;
 	
 	for (int i = 0, j = NUM_SEG-1; i < NUM_SEG; j = i++)
 	{
@@ -217,8 +217,8 @@ void duAppendCylinderWire(struct duDebugDraw* dd, float minx, float miny, float 
 	}
 }
 
-void duAppendBoxWire(struct duDebugDraw* dd, float minx, float miny, float minz,
-					 float maxx, float maxy, float maxz, unsigned int col)
+void duAppendBoxWire(struct duDebugDraw* dd, double minx, double miny, double minz,
+					 double maxx, double maxy, double maxz, unsigned int col)
 {
 	if (!dd) return;
 	// Top
@@ -252,8 +252,8 @@ void duAppendBoxWire(struct duDebugDraw* dd, float minx, float miny, float minz,
 	dd->vertex(minx, maxy, maxz, col);
 }
 
-void duAppendBoxPoints(struct duDebugDraw* dd, float minx, float miny, float minz,
-					   float maxx, float maxy, float maxz, unsigned int col)
+void duAppendBoxPoints(struct duDebugDraw* dd, double minx, double miny, double minz,
+					   double maxx, double maxy, double maxz, unsigned int col)
 {
 	if (!dd) return;
 	// Top
@@ -277,11 +277,11 @@ void duAppendBoxPoints(struct duDebugDraw* dd, float minx, float miny, float min
 	dd->vertex(minx, maxy, minz, col);
 }
 
-void duAppendBox(struct duDebugDraw* dd, float minx, float miny, float minz,
-				 float maxx, float maxy, float maxz, const unsigned int* fcol)
+void duAppendBox(struct duDebugDraw* dd, double minx, double miny, double minz,
+				 double maxx, double maxy, double maxz, const unsigned int* fcol)
 {
 	if (!dd) return;
-	const float verts[8*3] =
+	const double verts[8*3] =
 	{
 		minx, miny, minz,
 		maxx, miny, minz,
@@ -312,31 +312,31 @@ void duAppendBox(struct duDebugDraw* dd, float minx, float miny, float minz,
 	}
 }
 
-void duAppendCylinder(struct duDebugDraw* dd, float minx, float miny, float minz,
-					  float maxx, float maxy, float maxz, unsigned int col)
+void duAppendCylinder(struct duDebugDraw* dd, double minx, double miny, double minz,
+					  double maxx, double maxy, double maxz, unsigned int col)
 {
 	if (!dd) return;
 	
 	static const int NUM_SEG = 16;
-	static float dir[NUM_SEG*2];
+	static double dir[NUM_SEG*2];
 	static bool init = false;
 	if (!init)
 	{
 		init = true;
 		for (int i = 0; i < NUM_SEG; ++i)
 		{
-			const float a = (float)i/(float)NUM_SEG*DU_PI*2;
-			dir[i*2] = cosf(a);
-			dir[i*2+1] = sinf(a);
+			const double a = (double)i/(double)NUM_SEG*DU_PI*2;
+			dir[i*2] = cos(a);
+			dir[i*2+1] = sin(a);
 		}
 	}
 	
 	unsigned int col2 = duMultCol(col, 160);
 	
-	const float cx = (maxx + minx)/2;
-	const float cz = (maxz + minz)/2;
-	const float rx = (maxx - minx)/2;
-	const float rz = (maxz - minz)/2;
+	const double cx = (maxx + minx)/2;
+	const double cz = (maxz + minz)/2;
+	const double rx = (maxx - minx)/2;
+	const double rz = (maxz - minz)/2;
 
 	for (int i = 2; i < NUM_SEG; ++i)
 	{
@@ -365,9 +365,9 @@ void duAppendCylinder(struct duDebugDraw* dd, float minx, float miny, float minz
 }
 
 
-inline void evalArc(const float x0, const float y0, const float z0,
-					const float dx, const float dy, const float dz,
-					const float h, const float u, float* res)
+inline void evalArc(const double x0, const double y0, const double z0,
+					const double dx, const double dy, const double dz,
+					const double h, const double u, double* res)
 {
 	res[0] = x0 + dx * u;
 	res[1] = y0 + dy * u + h * (1-(u*2-1)*(u*2-1));
@@ -375,44 +375,44 @@ inline void evalArc(const float x0, const float y0, const float z0,
 }
 
 
-inline void vcross(float* dest, const float* v1, const float* v2)
+inline void vcross(double* dest, const double* v1, const double* v2)
 {
 	dest[0] = v1[1]*v2[2] - v1[2]*v2[1];
 	dest[1] = v1[2]*v2[0] - v1[0]*v2[2];
 	dest[2] = v1[0]*v2[1] - v1[1]*v2[0]; 
 }
 
-inline void vnormalize(float* v)
+inline void vnormalize(double* v)
 {
-	float d = 1.0f / sqrtf(v[0]*v[0] + v[1]*v[1] + v[2]*v[2]);
+	double d = 1.0 / sqrt(v[0]*v[0] + v[1]*v[1] + v[2]*v[2]);
 	v[0] *= d;
 	v[1] *= d;
 	v[2] *= d;
 }
 
-inline void vsub(float* dest, const float* v1, const float* v2)
+inline void vsub(double* dest, const double* v1, const double* v2)
 {
 	dest[0] = v1[0]-v2[0];
 	dest[1] = v1[1]-v2[1];
 	dest[2] = v1[2]-v2[2];
 }
 
-inline float vdistSqr(const float* v1, const float* v2)
+inline double vdistSqr(const double* v1, const double* v2)
 {
-	const float x = v1[0]-v2[0];
-	const float y = v1[1]-v2[1];
-	const float z = v1[2]-v2[2];
+	const double x = v1[0]-v2[0];
+	const double y = v1[1]-v2[1];
+	const double z = v1[2]-v2[2];
 	return x*x + y*y + z*z;
 }
 
 
-void appendArrowHead(struct duDebugDraw* dd, const float* p, const float* q,
-					 const float s, unsigned int col)
+void appendArrowHead(struct duDebugDraw* dd, const double* p, const double* q,
+					 const double s, unsigned int col)
 {
-	const float eps = 0.001f;
+	const double eps = 0.001;
 	if (!dd) return;
 	if (vdistSqr(p,q) < eps*eps) return;
-	float ax[3], ay[3] = {0,1,0}, az[3];
+	double ax[3], ay[3] = {0,1,0}, az[3];
 	vsub(az, q, p);
 	vnormalize(az);
 	vcross(ax, ay, az);
@@ -429,24 +429,24 @@ void appendArrowHead(struct duDebugDraw* dd, const float* p, const float* q,
 	
 }
 
-void duAppendArc(struct duDebugDraw* dd, const float x0, const float y0, const float z0,
-				 const float x1, const float y1, const float z1, const float h,
-				 const float as0, const float as1, unsigned int col)
+void duAppendArc(struct duDebugDraw* dd, const double x0, const double y0, const double z0,
+				 const double x1, const double y1, const double z1, const double h,
+				 const double as0, const double as1, unsigned int col)
 {
 	if (!dd) return;
 	static const int NUM_ARC_PTS = 8;
-	static const float PAD = 0.05f;
-	static const float ARC_PTS_SCALE = (1.0f-PAD*2) / (float)NUM_ARC_PTS;
-	const float dx = x1 - x0;
-	const float dy = y1 - y0;
-	const float dz = z1 - z0;
-	const float len = sqrtf(dx*dx + dy*dy + dz*dz);
-	float prev[3];
+	static const double PAD = 0.05;
+	static const double ARC_PTS_SCALE = (1.0-PAD*2) / (double)NUM_ARC_PTS;
+	const double dx = x1 - x0;
+	const double dy = y1 - y0;
+	const double dz = z1 - z0;
+	const double len = sqrt(dx*dx + dy*dy + dz*dz);
+	double prev[3];
 	evalArc(x0,y0,z0, dx,dy,dz, len*h, PAD, prev);
 	for (int i = 1; i <= NUM_ARC_PTS; ++i)
 	{
-		const float u = PAD + i * ARC_PTS_SCALE;
-		float pt[3];
+		const double u = PAD + i * ARC_PTS_SCALE;
+		double pt[3];
 		evalArc(x0,y0,z0, dx,dy,dz, len*h, u, pt);
 		dd->vertex(prev[0],prev[1],prev[2], col);
 		dd->vertex(pt[0],pt[1],pt[2], col);
@@ -454,26 +454,26 @@ void duAppendArc(struct duDebugDraw* dd, const float x0, const float y0, const f
 	}
 	
 	// End arrows
-	if (as0 > 0.001f)
+	if (as0 > 0.001)
 	{
-		float p[3], q[3];
+		double p[3], q[3];
 		evalArc(x0,y0,z0, dx,dy,dz, len*h, PAD, p);
-		evalArc(x0,y0,z0, dx,dy,dz, len*h, PAD+0.05f, q);
+		evalArc(x0,y0,z0, dx,dy,dz, len*h, PAD+0.05, q);
 		appendArrowHead(dd, p, q, as0, col);
 	}
 
-	if (as1 > 0.001f)
+	if (as1 > 0.001)
 	{
-		float p[3], q[3];
+		double p[3], q[3];
 		evalArc(x0,y0,z0, dx,dy,dz, len*h, 1-PAD, p);
-		evalArc(x0,y0,z0, dx,dy,dz, len*h, 1-(PAD+0.05f), q);
+		evalArc(x0,y0,z0, dx,dy,dz, len*h, 1-(PAD+0.05), q);
 		appendArrowHead(dd, p, q, as1, col);
 	}
 }
 
-void duAppendArrow(struct duDebugDraw* dd, const float x0, const float y0, const float z0,
-				   const float x1, const float y1, const float z1,
-				   const float as0, const float as1, unsigned int col)
+void duAppendArrow(struct duDebugDraw* dd, const double x0, const double y0, const double z0,
+				   const double x1, const double y1, const double z1,
+				   const double as0, const double as1, unsigned int col)
 {
 	if (!dd) return;
 
@@ -481,26 +481,26 @@ void duAppendArrow(struct duDebugDraw* dd, const float x0, const float y0, const
 	dd->vertex(x1,y1,z1, col);
 	
 	// End arrows
-	const float p[3] = {x0,y0,z0}, q[3] = {x1,y1,z1};
-	if (as0 > 0.001f)
+	const double p[3] = {x0,y0,z0}, q[3] = {x1,y1,z1};
+	if (as0 > 0.001)
 		appendArrowHead(dd, p, q, as0, col);
-	if (as1 > 0.001f)
+	if (as1 > 0.001)
 		appendArrowHead(dd, q, p, as1, col);
 }
 
-void duAppendCircle(struct duDebugDraw* dd, const float x, const float y, const float z,
-					const float r, unsigned int col)
+void duAppendCircle(struct duDebugDraw* dd, const double x, const double y, const double z,
+					const double r, unsigned int col)
 {
 	if (!dd) return;
 	static const int NUM_SEG = 40;
-	static float dir[40*2];
+	static double dir[40*2];
 	static bool init = false;
 	if (!init)
 	{
 		init = true;
 		for (int i = 0; i < NUM_SEG; ++i)
 		{
-			const float a = (float)i/(float)NUM_SEG*DU_PI*2;
+			const double a = (double)i/(double)NUM_SEG*DU_PI*2;
 			dir[i*2] = cosf(a);
 			dir[i*2+1] = sinf(a);
 		}
@@ -513,8 +513,8 @@ void duAppendCircle(struct duDebugDraw* dd, const float x, const float y, const 
 	}
 }
 
-void duAppendCross(struct duDebugDraw* dd, const float x, const float y, const float z,
-				   const float s, unsigned int col)
+void duAppendCross(struct duDebugDraw* dd, const double x, const double y, const double z,
+				   const double s, unsigned int col)
 {
 	if (!dd) return;
 	dd->vertex(x-s,y,z, col);
@@ -532,7 +532,7 @@ duDisplayList::duDisplayList(int cap) :
 	m_cap(0),
 	m_depthMask(true),
 	m_prim(DU_DRAW_LINES),
-	m_primSize(1.0f)
+	m_primSize(1.0)
 {
 	if (cap < 8)
 		cap = 8;
@@ -547,9 +547,9 @@ duDisplayList::~duDisplayList()
 
 void duDisplayList::resize(int cap)
 {
-	float* newPos = new float[cap*3];
+	double* newPos = new double[cap*3];
 	if (m_size)
-		memcpy(newPos, m_pos, sizeof(float)*3*m_size);
+		memcpy(newPos, m_pos, sizeof(double)*3*m_size);
 	delete [] m_pos;
 	m_pos = newPos;
 
@@ -572,18 +572,18 @@ void duDisplayList::depthMask(bool state)
 	m_depthMask = state;
 }
 
-void duDisplayList::begin(duDebugDrawPrimitives prim, float size)
+void duDisplayList::begin(duDebugDrawPrimitives prim, double size)
 {
 	clear();
 	m_prim = prim;
 	m_primSize = size;
 }
 
-void duDisplayList::vertex(const float x, const float y, const float z, unsigned int color)
+void duDisplayList::vertex(const double x, const double y, const double z, unsigned int color)
 {
 	if (m_size+1 >= m_cap)
 		resize(m_cap*2);
-	float* p = &m_pos[m_size*3];
+	double* p = &m_pos[m_size*3];
 	p[0] = x;
 	p[1] = y;
 	p[2] = z;
@@ -591,7 +591,7 @@ void duDisplayList::vertex(const float x, const float y, const float z, unsigned
 	m_size++;
 }
 
-void duDisplayList::vertex(const float* pos, unsigned int color)
+void duDisplayList::vertex(const double* pos, unsigned int color)
 {
 	vertex(pos[0],pos[1],pos[2],color);
 }
